slo_details: use float for hours

+++ analytics/dbt/models/spark/platform/slo_details.sql
@@ -9,5 +9,5 @@
 ) }}
 
 SELECT *,
-  cast(expected_time_utc_hours_str as int) as expected_time_utc_hours
+  cast(expected_time_utc_hours_str as float) as expected_time_utc_hours
 FROM {{ref("slo_details_seed")}}
